Add test signal by event
The test checks that a SIGIO signal is generated for each page fault. The
signal handler can itself trigger page faults, but these events are dropped
because the signal is blocked while executing. We also check that if we disable
the event, the page faults in the signal handler are not taken into account.

Depending on the situation, the number of page faults varies slightly, we thus
relax the assertion with a threshold of 5 events.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>